paged query

+++ src/main/java/ch/unil/fcrepo4/spring/data/core/query/FedoraQuery.java
@@ -9,7 +9,7 @@ public interface FedoraQuery {
 
     boolean isPaged();
 
-    int getOffset();
+    long getOffset();
 
-    int getLimit();
+    long getLimit();
 }
